+++ public/style.css
@@ -106,21 +106,31 @@ nav a {
 	margin-left: 0.25em;
 }
 
+@media screen and (max-width: 768px) {
+	.logo img {
+		width: 20px;
+		height: 20px;
+		margin-left: 0.5em;
+	}
+	.logo h3 {
+		font-size: 18px;
+	}
+	.search {
+		margin-left: 0;
+	}
+}
+
 @media screen and (max-width: 1254px) {
 	.logo {
 		margin-left: 0.5em;
 	}
 	.logo img {
-		width: 20px;
-		height: 20px;
 		margin-left: 0.25em;
 	}
 	.logo h3 {
-		font-size: 16px;
+		font-size: 18px;
 	}
-	.search {
-		margin-left: 0;
-	}
+
 	nav a {
 		display: inline-block;
 		margin: 0.25em;
@@ -224,6 +234,7 @@ form div {
 	object-fit: cover;
 	margin-bottom: 2em;
 }
+
 .products {
 	display: flex;
 	flex-wrap: wrap;
@@ -242,17 +253,33 @@ form div {
 	/* border: 1px solid rgb(150, 148, 148); */
 }
 
-@media screen and (max-width: 768px) {
+@media screen and (max-width: 878px) {
 	.products {
-		margin-left: 0;
+		margin-left: 15%;
 	}
 	.product {
 		max-width: calc(100% * (1 / 2));
 	}
 }
 
+.my-carousel {
+	height: 400px;
+	display: flex;
+	overflow: hidden;
+}
+
 .my-carousel img {
-	transition: opacity ease-in 0.5s;
+	transition: opacity ease-in-out 0.5s;
+}
+
+.hidden {
+	opacity: 0;
+	z-index: 0;
+}
+
+.visible {
+	opacity: 1;
+	z-index: 1;
 }
 
 .product-image {
